requirements.txt
Django==5.0.6
django-environ==0.11.2
stripe==10.6.0
gunicorn==22.0.0